requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-telegram-bot>=21.0
